WIP handling poorly-formatted Activity Insight data
The CSV dumps that we’re getting from Activity Insight have multiple
duplicate header names. This commit allows us to process these files
with the SmarterCSV library by doing some pre-processing to rename the
duplicate headers with unique names. We’ll still need to implement the
logic to verify that only one of the duplicate columns contains data for
each row. We also still need to implement the logic to map the CSV
columns (including the duplicates) to database columns. We’ll also
probably want at least a minimal component test for this functionality.