Create a model for storing individual imports of a publication
Conceptually, we need a place to store the canonical information for a 
given publication where we’ll keep track of the most accurate metadata
that we can gather/curate for that publication. We’ll also need a place 
to store information about the given publication that has been imported
from one or more sources (perhaps multiple times) so that we can allow
users to select the most accurate information if there are multiple,
differing sources.