Remove columns from publications database table
These columns have been added to the publication_imports table (with a
few changes). For now we don’t need to store any of this data directly
in a publication record - although that may change later. This breaks
the publications API and the publications importer. We need to make a
few more changes to our data model and then rethink the logic for how
a publication finds its data and how the data gets imported from CSV
files.