Make the Activity Insight user importer idempotent
As we’re instantiating new users from rows in the CSV file, we need to
check to see if a user with the same WebAccess ID already exists in the
database. We also need to check whether or not there is more than one
row with the same access ID within a given batch since the batches are
bulk-inserted and model validations are bypassed.