Remove authentication from the “public” home page
The only reason we were authenticating users on the home page was to
allow us to show different navigation depending on the type of user at
some point later on. However, it seems that the VM Hosting load
balancer’s health checks for our application server is failing because
this is causing it to receive a 401 response for the home page. I’m
going to try removing this to see if we can get our production
environment working. If so, it might just be easier to have an actual
“public” landing page that doesn’t require authenticating via WebAccess.
For now, the whole website is behind the VPN anyway.