Initialize the swagger ui validator
This is supposed to validate our API's JSON file against the Swagger 2.0 spec. Setting the validator URL to nil seems wrong, but the URL it constructs with this setting actually looks correct. On the staging (our beta) site it looks like:

http://online.swagger.io/validator?url=https://stage.metadata.libraries.psu.edu/api_docs/swagger_docs/v1/swagger.json

When our API spec is valid Swagger 2.0 a green "Valid {...}" badge should appear in the bottom right of the footer on the api docs page. NOTE: in development, the error badge, red "ERROR {...}", will always be displayed since the validator can't read the JSON file from localhost.