Declare permitted admin actions individually
There are a lot of admin actions that we want to restrict in order to
allow admins to do only what they need to do, keep the admin interface
from becoming unnecessarily cluttered, and prevent admin users from 
changing data in ways that we don’t expect. However, it’s a little too
restrictive to ignore any of our models entirely, since it prevents
admins from viewing some data that might be relevant.