Remove ‘institution’ attribute from users
We’re not currently storing any meaningful data in this column, so it
seems better to just remove it for now. We may want to attempt to model
the organizational structure of Penn State at some point and associate
users with various institutes/groups/departments/colleges/etc., but we
wouldn’t be able to do that using this single user attribute anyway.