Don’t update manually-edited publications during imports
This also improves the overall component test coverage of the Pure
publication importer. Coverage of the importers for Activity Insight
publication data is still lacking because I’ve had difficulty setting up
the fixtures that would be required for component testing. Some
refactoring and unit testing around all of the importer classes is still
badly needed.