Remove ‘title’ attribute from users
We’re not currently importing any data into this field, and this
probably wouldn’t be the way that we’d actually want to store such
data anyway. A user’s title, if we ever have a use case for modeling it,
will probably be a more complicated concept that involves the user’s 
past and present roles in multiple organizations of various types.