Remove null constraint and validation on presentation titles
Many records that we’re importing do not have any value for `title`.
There is also a `name` attribute for presentations, but in a few cases,
it looks like there is no value for either `name` or `title`. It’s
probably best to import all of the data, regardless of missing values,
and sort it out later.