Fix bug in Activity Insight CSV importing
Failing to close the tempfile for writing before allowing the CSV parser
to open it for reading was causing the parser to unexpectedly hit the
end of the file. It seems like this was failing silently (or we were
just missing it somehow) during real imports with larger CSV files, and
we ended up not quite importing all of the rows as a result. Since the
larger files are processed in batches, I’m guessing that perhaps all of
the batches except for the last one were being imported successfully.

I happened to catch this because an obvious error was thrown during
testing with a small fixture file that had many less rows than our
default batch size.