Temporarily disable API authentication
This will allow us to merge and deploy part of the functionality of this
feature without interrupting access for current users of the API before
we’ve had a chance to issue API keys. With this commit deployed, we can
create API keys in the production database and issue them to users, and
users can see the Swagger documentation for the API authentication. When
all of the keys are issued, we can revert this commit and deploy to turn
on authentication.