Only list performances that are unique as displayed
We’re only displaying certain attributes of each performance for the
user profile. In some cases, users have two or more performances where
these attributes are identical, so on a profile page they appear to be
duplicates even though they are not. We looked at displaying additional
attributes to differentiate them instead, but the data was too messy.