Import Activity Insight “profile” data separately from users
In the past, this data was combined with another export of other user
data from Activity Insight before we imported it. This was sub-optimal
because of the manual effort required to combine the exports and because
some of the data was getting truncated as a result of the combination.
We’re now going to move toward importing each of these Activity Insight
exports separately, starting here with the data that had been truncated.