Update fixture to reflect new scenario in real Pure data
Up until now, every person object that we received from Pure had a top-
level attribute called `externalId`, and the value of that attribute
was that person’s Penn State Webaccess ID (which is how we identify
users in this app). We just downloaded some new data, and at least one
person object was completely missing this attribute, so at the very
least, we cannot depend on the attribute’s presence for our importer to
work properly. It appears that most person objects do still have this
attribute as expected, so for now we’ll probably just ignore objects
where it’s missing. However, it’s worth noting that we don’t really know
why this attribute may be missing in some cases, and that might call 
into question its reliability for our use case.