Always import publication DOIs from Pure
This is a temporary change so that we can update some of the data in
production. In the production database, we had many publications
imported from Pure, and none of them had DOI values since we only
started importing those recently. Further, many of those publications
had been flagged as having been edited by an admin user because they
were merged with other duplicate publication records. This flag
prevented the importer from importing DOI values for these records.
We’ll revert this behavior once we’ve imported DOIs for all of the
existing Pure publications in the production database.