Import DOI values from Activity Insight publication data
A couple of things to note here:

As with importing DOI data from Pure, we’re temporarily going to update
those values for existing records, even if those records have been
modified by an admin user. This will allow us to import all available
DOI data for publications in the production database. After we’ve done
that, we’ll revert this behavior so that AI imports won’t update any
data on admin-modified records again.

Also, I started to write the tests for ActivityInsightPublication before
I realized that I wanted to extract the DOI parsing logic, so these
tests, although they are in spec/unit are actually written as component
tests now, and with a little more coverage than they really need. This
could be refactored, but it’s not too bad as-is.