Return all matching publications instead of the first one found
Since we’ll be attaching grants to existing publications from other
import sources, we may encounter duplicate publications that have not
yet been merged. All duplicates of a given publication should have the
grants attached so that those associations persist even if one or more
duplicate publications is deleted as the result of a merge.