Look for DOIs in another location in Web of Science data
Some WoS publications have an identifier element with a “doi” type, and
other have one with an “xref_doi” type. I’m not sure what, if any,
difference there is semantically between these, but they both seem to
contain values that are DOIs for the given publication, so it seems like
we should be using data from whichever one is present.