Convert unit spec to a component spec
This class still has all of its dependencies mocked out, but since we’re
stubbing methods on an existing ActiveRecord class, it works better to
load the actual class rather than redefining it this spec. That way the
stubbed class definition that we previously had here doesn’t interfere
with the other tests when they’re all loaded and run together.