Start parsing out canonical agency names from Web of Science grant data
For now, we’re just interested in identifying NSF grants since we’ll be
importing additional grant data directly from NSF soon. This algorithm
may also be imperfect, but it’s based on a pretty large sample of data,
so it should at least identify the vast majority of NSF grants.