Remove access to contract data from API
Contracts were entities that were imported from Activity Insight which
included research grants. We decided that we don’t have a good way of
knowing what contract data would be ok to provide via our API and what
data is somewhat private/sensitive. So we’ve abandoned the effort to
acquire this data from Activity Insight, curate it, and provide it to 
our API consumer in favor of importing grant data that is already 
publicly available from sources like Web of Science, the National
Science Foundation, etc.

While this commit just removes the contracts endpoint from the API (and
supporting code), we will probably end up dropping the data from the
production database and then removing all of the code for importing,
storing, and managing it from the application.

I left the spot in the user profile API where we formerly provided
contract data as a placeholder for the public grant data that we’ll be
adding in the future.