Remove unused routes
The `resources` method declares a bunch of extra routes that we haven’t
implemented and don’t intend to implement. Unless we’re actually
implementing all of the routes that `resources` creates, we should 
instead declare only the routes that we *are* implementing explicitly 
one by one. This is cleaner, more correct, and it leaves less holes on
the surface of the application.