Add an Open Access Button check timestamp to publications
Since the import of data from Open Access Button takes such a long time
and has the potential to be interrupted, we want to be able to pick up
where we left off if we stop in the middle of an import for any reason.
Starting from the beginning would waste time since it’s unlikely that
any of the data would have changed since the start of the original
import.