Add branding to the profile management page
This change tries to match up the visual style and branding between the
public profile page in the metadata app and the profile management
interface. This has been done without actually reusing the same template
and styling code because the styling for the public page was taken from
the profile demo site which was designed by a 3rd party and which uses
a different CSS framework than what we’re using for the rest of the app.