Prohibit publication editing after submitting to ScholarSphere
If a publication author has gone through the workflow to submit their
publication to ScholarSphere to make it open access, then we’re waiting
for a metadata admin to look in ScholarSphere for their submission,
harvest the URL, and enter it into the metadata database. In the
meantime, the author shouldn’t be able to modify the publication or go
through the same workflow again, and neither should any other author on
the publication.