Change behavior for authentication failure due to missing user record
For a couple of specific URLs, we wanted to redirect to a particular
page in the case where a user had successfully authenticated to Penn 
State WebAccess but didn’t have a user record in the database. What we 
were doing here before didn’t work right anyway. The authentication 
header comes from Cosign, so there’s no way for the user to get back to
the Rails app after having entered an invalid “WebAccess ID or 
password”, and the only thing that gets passed back to the Rails app is
a username when the user successfully authenticates to WebAccess. The
result was that if a WebAccess-authenticated user without a user record
in the database tried to access an authenticated page, Devise’s Warden
failure app would just redirect them back to the new session action,
which in turn would redirect them back to the Cosign/WebAccess login
which then created a redirect loop.