Allow admin users to update the open access URL on certain publications
This allows the semi-manual workflow for depositing open access
publications in ScholarSphere to be completed. Now an admin user can see
a list of authorships for which the authors have followed the link in
their metadata profile to upload an open access copy of the publication
to ScholarSphere. The admin can then manually look for that publication
in ScholarSphere, find the URL to the open access content, and record
that URL in the corresponding publication record in the metadata
database using the form that is being created here.