Add some XML test fixture data for Activity Insight publications
We’ll be implementing new Activity Insight import functionality that
imports publications, authorships, and contributors via Activity
Insight’s API, and we’ll use this fixture data for testing. For better
test coverage, we may need to add a little more data and a few more
examples to this - i.e. to cover cases where the contribution type
matches some other variant of “journal article” or where journal name or
publisher has a value of “other”.