Allow admins to edit the open access URL for any publication
This is already possible in the workflow where admins add open access
URLs for publications that have been added to Scholarsphere, but there
are other manual workflows whereby a publication may be made open
access, and an admin would then need to add that URL to the 
publication’s record in the metadata database. So it’s important that an
admin can edit the open access URL for any publication along with the
rest of the publication’s metadata on the main “edit publication” page.