Fix bug in open access workflow
If an admin user creates or updates a publication, the open access
fields may be saved with blank strings rather than null values. We want
to treat any publications with blank values the same as those with null
values. It might ultimately be better to convert blank strings to nulls
when the publication is saved, but since we’re using an admin framework, that would probably be a little tricky/messy to do.