Hide navigation to “bio” section of profile management interface
I’m going to merge the sandbox ORCiD integration into master. As a
result, it will be deployed to production. That won’t be a problem, but
the ORCiD integration won’t work in production yet, and we don’t want
people to see it, so we’re just going to hide the link to it for now.