Track when an employment record has been added to ORCID
This will prevent a user from accidentally adding the same record more
than once, and it will allow us to enable the user to update or delete
the record later. We don’t currently have a great way of knowing when
a tracked record has been deleted from ORCID via the ORCID website, so
things could get out of sync. We could solve this problem with another
API call or possibly with an ORCID web hook.