Add columns for other ORCID authentication/authorization info to users
As per ORCID’s guidelines, we need to store this information as long as
we’re implementing our own, independent Oauth workflow, and we may need 
to use this information later.

Users already had an `orcid_identifier` attribute that we’re currently
populating from LDAP data where it’s available. In most cases, this
should be the same ORCID iD that we get back when the user authenticates
to ORCID through our app, but it’s possible that it could be different
if a user has two ORCID accounts and they try to link a different
account to their metadata profile than the one that they linked to their
Penn State access account. So as long as we’re doing ORCID Oauth
independently of other apps/systems at Penn State, we should probably
store and use the ORCID iD that we get back from our own ORCID Oauth
interaction.

We can still use the presence of the ORCID iD from LDAP as the flag that
determines whether or not a user can link their metadata profile to
their ORCID record. That should help to encourage people to link their
Penn State access account to their ORCID iD, which is how we may
prefer to authenticate them in the future if that becomes possible.