Use the iD that we got directly from ORCID in API calls
The ORCID iD value that we had be using is gotten indirectly through
LDAP and as a result of a separate Oauth process. Hence, it’s possible
that it could be different from the value that we’ve obtained directly
(which is the value that is definitely associated with the access token
that we also obtained directly).

See the message from commit 44df96e4d45adaf75994911b3d45927c3922c4fb.