Temporarily use flash message to understand ORCID OAuth responses
ORCID’s documentation tells me that if a user denies authorization for
our app, then ORCID will send us an error response/message, but it
doesn’t give me any details as to what that response looks like. I’m
trying to figure this out because ideally, we should redirect the user
to a different page in our app if they deny authorization.