Change entry point for ORCID OAuth “authorization denied” workflow
When a user denies OAuth authorization for our app, ORCID actually
redirects the user back to us with a specific error message in the query
string params. This is the right way to detect that the user clicked
“Deny” on the ORCID site and send them into the proper workflow in our
app. Any failure to actually get an access token (assuming that the
“access denied” error isn’t present) can be handled separately.