Find the set of users who should get an open access notification
Admittedly, the test for this is nuts, but the logic that it’s testing
is really complicated too, and we want to be sure that only the correct
users will get emailed. I’m not sure if there’s a great way to
encapsulate any of this logic somewhere else, but even if there is, I’m
not sure that we’d want to. I imagine that this whole method could
somehow be expressed as a single database query, but I’m probably not a
good enough SQL ninja to do that. If I could though, it would probably
be a lot better in terms of performance. So having all of this logic in
one method and having this ridiculous test for it might allow me to come
back later and attempt to refactor this.