Rewrite ridiculously long test
While the previous test wasn’t that difficult to produce, it was pretty
much impossible to comprehend or modify. It was more thorough in that
it tested every permutation of relevant state, but many of those
permutations shouldn’t ever really happen in practice. Moreover, the
algorithm that we’re testing is basically taking a base set of data and
applying a series of filters to it. Ensuring that each filter has been
applied correctly is the most important aspect of testing this.