Address warning from ruby
The interpreter kept producing a warning about the `:degree` key in
this argument being overwritten since it is used twice. In spite of
the warning and the fact that the overwritten key seems like it should
cause the method to behave incorrectly, the method actually seemed to
work correctly and the tests passed. So this change should just be a
refactoring that gets rid of the warning.