Record when an open access notification is sent for each authorship
We already store this information in the user record which is our basis
to determine frequency for email sending for a given person. We’re just
going to use presence of the timestamps on the authorships to partition
the list of publications in the notification email into those 
publications that the user has been notified about before, and those
that they haven’t been notified about before.