WIP on refining duplicate publication detection
Our original method of comparing publications in order to find
duplicates was much too strict, and it missed a lot of actual
duplicates. This change seems to find a much higher percentage of the
actual duplicates without producing many false-positive matches. This
could probably use a few more test cases, and we’ll need to evaluate the
results after this has been used on a large set of real data to
determine whether or not it needs further refinement.