WIP on updating duplicate publication group workflows
This implements the new workflow that we need for merging publications
and deleting duplicate groups, but it’s missing some integration tests,
and some of the old integration tests haven’t been updated so they’re
still failing. Also included is a stub for the “ignore false duplicates”
workflow, but this workflow hasn’t been implemented yet.