Ignore known false-positive matches when grouping duplicate publications
I think this is correct, but it was pretty tricky to wrap my head
around, so I’m not completely convinced that it works correctly under
all circumstances without having some additional test cases, but I also
can’t think of other meaningful cases to test right now.

This has also revealed that there’s still more work to do elsewhere to
ensure that the data stays consistent. For example, two publications
that are already in the same non-duplicate group could also end up in the
same duplicate group if a new, third publication matches both of them.
In this case, we should not allow publications that are in the same
non-duplicate group to be merged. Also, when a publication is merged
with another and then deleted, we’ll need to replace the publication that
we’re deleting with the publication that we’re keeping in each of the
non-duplicate groups to which the publication to be deleted belongs.
That way we won’t lose the information about the false-positive matches
of the publication that we’re deleting.