Remove DOI from duplicate publication comparison
This was assuming that the DOI data was always comparable if it was
present, but that’s not actually the case. DOI data from Pure is
consistent and comparable, but DOI data from Activity Insight is very
inconsistent in formatting and even in meaning. Faculty users can
apparently enter any value at all into this field when they’re entering
their publication data in Activity Insight, so the value is sometimes
formatted as a URL, sometimes not, and sometimes it’s present but it’s
not a DOI at all.