Get the new organization/publications query tests to pass
The lines that are being deleted here were just scratch code from the
previous WIP commit.

Some explanation for what we’re aiming to do here:

We need the application to be able to answer queries such as, “show me
all of the publications that are associated with a given organization”.
The trick is that publications are associated with organizations through
a chain of associations (publication -> authorship -> user ->
user_organization_membership -> organization) and organizations are
associated with each other in a tree of parent/child relationships. When
we ask for the publications for a given organization, what we really
want is all of the publications that were authored by members of that
organization and any of its descendant organizations during those
memberships.

When we import organization and membership data, the source data only
indicates membership in a single organization, and it doesn’t explicitly
indicate membership in any of the other organizations from which that
organization descends in the tree - even though memberships in those
organizations are implied (if you’re explicitly a member of Biology
Department, then you’re implicitly a member of Penn State University
even if the source data doesn’t say so).

To solve the problem of getting of publications for all descendent
organizations of a given organization, we had been explicitly creating
all of those implicit memberships whenever we imported the source data.
That made the queries trivial, but there were a few problems with this.
First, creating those implicit memberships was buggy. For example, 
sometimes implicit memberships would overlap with each other, and we’d
have no good way of identifying when that was happening, so we’d fail
to create some of the memberships, and thus we’d end up missing some
publications when we ran the queries. Another problem is that it is
difficult to keep all of those implicit memberships up to date correctly
whenever we import new and/or updated source data. Likewise, whereas
we only had one source of organization membership data in the past, we
now need to import from a second source which would also require us to
create implicit membership records and somehow reconcile them with
records for the other source (knowing that there will also likely be
duplication of data between the sources).

A much better solution to this same problem seems to be to query through
the explicit organization memberships recursively rather than
recursively creating all of the implicit memberships in order to 
simplify the querying. So that’s what we’re trying to do here. Once we
can do essentially the same querying using only the explicit membership
data that we’ve imported, we can stop creating new implicit memberships
and delete those that we’ve already created. That will greatly simplify
the importing of our second source of membership data.