WIP on a custom Rails Admin action to list publications by organization
Since Rails Admin and the underlying ActiveRecord ORM don’t support the
recursive querying that we need to do in order to correctly list all of
the publications for a given organization, we can’t take advantage of
all of Rails Admin’s out-of-the-box functionality without doing a lot of
hacking around way down in its guts.

There’s more that we need to do here. 

First of all, the stuff that has already been added needs a pretty
thorough integration test. We’ve copied code out of the Rails Admin
framework as a basis for a customized action and view. This code has
dependencies on Rails Admin, so future updates to the framework may very
well break it.

Second, We skirted around authorization for this custom action, and I’m
not sure if that’s a problem yet (or if it will be in the future). We
might need to reimplement authorization in the same way that the
framework does.

Finally, it will be important for admin users of the app to be able to
export data using the scope defined for this new custom action, but we
don’t get that for free - we’ll have to customize the export
functionality as well.