Check for existing data when importing organization memberships from Pure
We’re going to be importing the same kind of data from one of Penn
State’s HR systems, and it could overlap with data that we import from
Pure. When that happens, we don’t want to create records that may be
duplicates, so we’ll check for records that seem to be the same, and
we’ll take the Pure data over the HR data.

I should note that I’m not sure if there is actually semantic
equivalence between all of the data that we’re getting from these two
different sources. In particular, since we’re importing the
organizations themselves from Pure, we’re kind of guessing when it comes
to matching those organizations to what is in the HR data. I’m also not
totally certain that the start dates that we’re getting from each source
actually mean the same thing. For now, it’s more important just that we
have all of this data even if we’re unaware that it may have somewhat
different meaning depending on where it was imported from.