Import user metadata from an export generated by a Penn State HR system
We were given a CSV file of user data that was exported from HR software
used by Penn State. I’m not sure how much (if any) the file was
manipulated after export and before it was given to us, but since it
only contained a couple of hundred records, I did some manual
manipulation to it before writing this importer so that the importer
could be simpler.

I’m not sure how repeatable it will be to import more data like this in
the future using the same importer, or whether or not we’ll even need to
do so again.

Obviously, this could use a little more testing, but probably nothing
too thorough considering that this might just be one-time-use code.