Land on the ‘Bio’ page of the profile management interface
This is a more natural starting point, and it’s the only page in the
profile management interface where a user can link their profile to
their ORCID record (which we want them to do, and which they need to do
in order to push data on some of the other profile management pages to
ORCID). We didn’t do this initially because the ‘Bio’ page was created
after the rest of the profile management interface.