Fix bug in importer
This is possible bug that was revealed in testing with realistic fixture
data. I’m not sure if it behaves the same way with real data, but there
appears to be some problem with the way that nokogiri parses the `& `
character (which is not escaped) in the fixture data. The resulting
text from the nokogiri parsing matches a single whitespace character
instead of `& `.