Further specify admin configuration for editing authorships
Originally we didn’t allow the editing of authorship records directly,
but we did allow authorship data to be edited from the edit page for
publications. Since we didn’t specify any admin configuration for
editing authorships, it defaulted to showing all of the authorship
fields (and allowing them to be edited) via the publication edit page.
Recently, we did begin to allow authorships to be edited directly, and
we specified a configuration for the edit action that only shows one
field from the authorship record. This resulted in a regression in an
admin’s previous ability to view/edit authorship information from the
publication edit page.

Here we’re explicitly reintroducing the other authorship fields to the
edit views for authorships. We’re not allowing all of those fields to be
edited (as was possible before) because we haven’t identified an actual
use case for that, but there is a use case for deleting authorships from
a publication. To be able to easily delete the correct authorship, an
admin at least needs to be able to see the user to which the authorship
belongs.