Add auto_merged flag to publication imports
We’re going to attempt to automatically merge certain duplicate
publication groups. Without a means to keep track of which imports were
automatically merged into other publications, it would be impossible to
undo what will be a large number of automatic merges. We know that we’ll
probably end up automatically merging some publications that aren’t
actually duplicates of each other. That in turn would delete unique
publication metadata from the database and prevent it from being
replaced. With that in mind, we should provide ourselves a way to
distinguish automatic merges from merges that were performed by a human
admin so that we can undo the automatic merges in the event that we
determine that too much data was lost in the process.