Make the open access notifier more flexible
The notifier should only ever send emails to users who need them, so
instead of notifying every user in any arbitrary collection of users
passed in by the caller, we should pick only the users that need a
notification from the collection that is passed in. This makes it easy
to notify all users that need notifications or to notify only a subset
of all users who need notifications.