Add a task for sending open access emails to PSU Libraries users
Prior to sending the very first round of these email notifications to
all of the applicable faculty at Penn State, we want to send emails just
to people within the Penn State University Park Libraries organization
so that they can evaluate how it’s working. This is just a convenience
task for making that happen.