Rename publisher attribute on publications to publisher_name
We’ve created a separate model to store publisher data that is going to
be associated with publications. That association is going to be called
`publisher`, so we need to give this text field another name. It’s not
clear yet if or how we’ll use this renamed field in the future, but
most of our publication data sources represent a journal’s publisher as
a simple string, so we’ll still need to store that information somehow.