Fix PSU Law school publication importing
We weren’t quite querying the law school repositories the way that we
should have been, and so we weren’t getting some of the publication
metadata that we wanted like the titles of the journals in which
articles were published. We also discovered a better way to refine the
query to get only the records that we want from the repositories.