Group and hide duplicates when importing new publications from Pure
It may often be the case that metadata for a new publication that is
being imported from Pure has already been imported from another source
such as Activity Insight. In that case, we want to group the new Pure
publication with other existing suspected duplicates as soon as it’s
created so that it’s queued up for manual deduplication later. We also
don’t want the duplicate data to be visible to end users, and we take
Pure’s metadata for publications to be canonical when there are known or
possible duplicates, so we’ll automatically hide any suspected
duplicates that we find. If they are truly duplicates, they’ll most
likely be deleted later anyway, and if they’re not, then they’ll be
manually unhidden as part of the deduplication process.