Don’t send open access email notifications for hidden publications
An admin user of the app can hide a publication for any reason, but
generally publications are hidden have been hidden automatically because
suspected duplicates of the publication have been identified and have
not yet been resolved. We don’t want to send an open access email to a
user if all that they have are publications that are hidden because
they’re likely duplicates and publications that otherwise don’t qualify
for an open access reminder email. If a publication is hidden because
it may be a duplicate, then eventually it will either be deleted or
unhidden. In the latter case, an open access reminder email will
eventually be sent for it.