WIP on importing publication-journal associations from Pure
This is just a work in progress at this point. Obviously, the changes
here are untested, so those tests need to be added. Some other things
will also need to be done as a result of these changes:

- We’ll need to further modify the Pure publication importer, at least
  temporarily, so that we get a complete import of these associations
  even for publication records that have been modified by an admin user.
- We’ll need to consider and possibly change the way in which we
  publicly present (via the API, the profile management interface, etc.)
  the names of a publication’s journal and publisher since we now have
  two possible sources for each of these names - the `journal_title` and
  `publisher_name` text fields on the publication itself (which are
  still important), and the names obtained through the journal and
  publisher associations. Most likely we’ll want to get these names from
  the associated records if they exist, and fall back to the text fields
  on the publication record if not.

We also still need a way to present the most important view of this
association data in the admin interface - which is a list of publishers
ordered by and including the number of publications that are associated
with each.