Modify pure publication import behavior
We think that we probably lost some publication DOIs in production that
were originally part of the metadata that we imported from Pure. The
values may have been lost during duplicate publication merges. This 
change will allow us to reimport the DOI values from the Pure metadata
for any publication records that are associated with a Pure import and
are missing a DOI - even if the records have been modified by an RMD
admin (i.e. merged with duplicates). We’ll *only* reimport a DOI if the
record doesn’t already have one; that way any DOIs that were manually
entered/fixed by an RMD admin will still be preserved.