Allow admin users to create and edit open access waivers
Faculty users sometimes ask for an open access publication waiver 
outside of the workflow of the RMD app, and this change allows RMD
admins to create waivers on behalf of these users. It will also allow
RMD admins to generally provide better support with regard to submitting
and correcting waivers.