Add a column to publications to store URLs to works in Scholarsphere
We need a place to store URLs to publications that have been deposited
in Scholarsphere via RMD. The two fields that we currently have,
`open_access_url` and `user_submitted_open_access_url` aren’t sufficient
for a few reasons. The Scholarsphere open access URLs will be entered
into RMD by admins for now and possibly by Scholarsphere itself someday,
and we don’t want to store them in the `user_submitted_open_access_url`
field because we may allow faculty users to edit the contents of that
field at some point since they are the users who would normally populate
it in the first place. If a Scholarsphere URL has been entered into the
publication record, we wouldn’t want faculty members to be able to edit
that URL. We don’t want to store Scholarsphere URLs in the
`open_access_url` field either because that’s where we store URLs that
have been harvested from Open Access Button, and we want to be able to
differentiate between publications that have become open access as a
result of Open Access Button and publications that have become open
access as a result of RMD/Scholarsphere. Also, if we happen to know
about open access copies of the same publication in multiple places on
the web including Scholarsphere, then we’d prefer to show the
Scholarsphere URL publicly in RMD instead of the other locations.